move & rename create-session-view
